sales_sql_app/requirements.txt
streamlit
langchain
langchain-community
langchain-openai
faiss-cpu
pandas
plotly
python-dotenv
sqlalchemy
openai
tiktoken
langchain-google-genai
putergenai
sentence-transformers
torch
transformers
accelerate
langchain-ollama

